Hide the huge title input
In the preview dropdown, there is an option to hide the template content.
In such case, the editor switches to post-only mode and displays the big
title input on top. Such a title makes no sense in an email context as it
is not part of the final render.

This commit adds CSS to hide it.
WOOPLUG-3951