Cleanup SCSS files and fix CSS loading
The CSS loading issue was caused by content related CSS for richtext defined in CSS which is used for layout
and was added additionally to the iframe.

I copied styles that we need for richtext to content-editor.css, which
is designated for CSS that we want to apply to content.

WOOPLUG-3951